Update config doc to show configuration is optional
Based on StackOverflow comment [1], documentation may imply that
configuration is required. Have edited this to stress that it is optional.

Also added a separate section for Code analysis rule set files instructions
so that it stands out in the same way as the nsubstitute.json option
(otherwise is easy to skim the docs and not find how to set up the rule
set files).

[1]: https://stackoverflow.com/questions/54196778/nsubstitute-receivedcallsexception-error-when-run-calls-in-a-sequence?noredirect=1#comment95318298_54196778